Refactor files in MVP fashion
