+++ app/src/androidTest/java/com/github/geohunt/app/MainActivityTest.java
@@ -6,16 +6,14 @@ import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.matcher.ViewMatchers.*;
 import static org.hamcrest.Matchers.allOf;
 
-import androidx.test.espresso.ViewInteraction;
-import androidx.test.espresso.action.ViewActions;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.intent.matcher.IntentMatchers;
-import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-import androidx.test.espresso.Espresso;
 
-import org.hamcrest.Matcher;
+import com.github.geohunt.app.view.GreetingActivity;
+import com.github.geohunt.app.view.MainActivity;
+
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
