+++ app/src/main/res/layout/activity_greeting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GreetingActivity">
+    tools:context=".view.GreetingActivity">
 
     <TextView
         android:id="@+id/greetingMessage"
